Update README.md
Agrihelp help is a react-based application. This website
was made for the welfare of farmers. It shows weather
forecast temperature, cloud %age, humidity percentage,
etc.
It predicts crop fire-catching probability using logistic
regression by using the current temperature, humidity. It
also suggests methods to prevent parali. There is also a
complaint form in the application for the farmers.